Move session helpers into a service class
